[v3.8.6] Use array.forEach instead of `for of` to optimize code size  (#18112)

+++ cocos/core/data/garbage-collection.ts
@@ -69,6 +69,7 @@ class GarbageCollectionManager {
     }
 
     private finalizationRegistryCallback (token: any): void {
+        if (!EDITOR) return;
         const gcObject = this._gcObjects.get(token);
         if (gcObject) {
             this._gcObjects.delete(token);
